test: BoardServiceTest 게시글_작성자_불일치로_삭제_실패 -> 테스트 성공

+++ src/main/java/com/hanghae/prevstudy/domain/board/service/BoardServiceImpl.java
@@ -129,10 +129,14 @@ public class BoardServiceImpl implements BoardService {
     }
 
     @Override
-    public void delete(Long boardId) {
+    @Transactional
+    public void delete(Long boardId, UserDetailsImpl userDetails) {
         Board board = boardRepository.findById(boardId)
                 .orElseThrow(() -> new PrevStudyException(BoardErrorCode.BOARD_NOT_FOUND));
 
+        if (!isWriterMatch(userDetails.getId(), board.getWriter().getId())) {
+            throw new PrevStudyException(BoardErrorCode.FORBIDDEN_ACCESS);
+        }
         boardRepository.delete(board);
     }
 }
